Extract FakeLogger into separate file
