+++ test/acceptance/fake_logger.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class FakeLogger
+  attr_reader :warnings
+
+  def initialize
+    @warnings = []
+  end
+
+  def warning(message)
+    @warnings << message
+  end
+end
